requirements.txt
llama-cpp-python
mistral_inference
diffusers 
transformers 
accelerate 
safetensors